Safari-609.1.11.1/LayoutTests/media/video-controller-child-rate-expected.txt


Test that setting playbackRate on video with controller is ignored.

EVENT(canplaythrough)
EVENT(playing)
END OF TEST
